Succescfully updated investmentblock
Simple case with low error
Changed configuration with p_nom, maxpower, etc.

+++ test/main.py
@@ -58,56 +58,56 @@ print(f"La classe di trasformazione ci mette {datetime.now() - then} secondi")
 
 tran = transformation.convert_to_blocks()
 
-### UCBlock configuration ###
-configfile = pysmspp.SMSConfig(template="uc_solverconfig")  # load a default config file [highs solver]
-temporary_smspp_file = "output/temp_network.nc"  # path to temporary SMS++ file
-output_file = "output/temp_log_file.txt"  # path to the output file (optional)
-solution_file = "output/temp_solution_file.nc"
+if transformation.dimensions['InvestmentBlock']['NumAssets'] == 0:
+    ### UCBlock configuration ###
+    configfile = pysmspp.SMSConfig(template="uc_solverconfig")  # load a default config file [highs solver]
+    temporary_smspp_file = "output/temp_network.nc"  # path to temporary SMS++ file
+    output_file = "output/temp_log_file.txt"  # path to the output file (optional)
+    solution_file = "output/temp_solution_file.nc"
+    
+    # Check if the file exists
+    if os.path.exists(solution_file):
+        os.remove(solution_file)
+    
+    result = tran.optimize(configfile, temporary_smspp_file, output_file, solution_file)
+    
+    statistics = network.statistics()
+    operational_cost = statistics['Operational Expenditure'].sum()
+    error = (operational_cost - result.objective_value) / operational_cost * 100
+    
+    print(f"Error PyPSA-SMS++ of {error}%")
+    
+    # Esegui la funzione sul file di testo
+    data_dict = parse_txt_file(output_file)
 
-# Check if the file exists
-if os.path.exists(solution_file):
-    os.remove(solution_file)
-
-result = tran.optimize(configfile, temporary_smspp_file, output_file, solution_file)
-
-### InvestmentBlock configuration ###
-
-# configfile = pysmspp.SMSConfig(template="InvestmentBlock/BSPar.txt")
-# temporary_smspp_file = "output/temp_network_investment.nc"
-# output_file = "output/temp_log_file_investment.txt"  # path to the output file (optional)
-# # BSC_NAME = "BSPar.txt"          # Name of the file describing the BlockSolverConfig
-
-# result = tran.optimize(configfile, temporary_smspp_file, output_file, inner_block_name='InvestmentBlock')
-
-
-statistics = network.statistics()
-operational_cost = statistics['Operational Expenditure'].sum()
-error = (operational_cost - result.objective_value) / operational_cost * 100
-
-# objective_pypsa = network.objective + network.objective_constant
-# objective_smspp = result.objective_value
-# error = (objective_pypsa - objective_smspp) / objective_pypsa
-
-print(f"Error PyPSA-SMS++ of {error}%")
+    print(f"Il solver ci ha messo {data_dict['elapsed_time']}s")
+    print(f"Il tempo totale (trasformazione+pysmspp+ottimizzazione smspp) è {datetime.now() - then}")
 
 
+    solution = transformation.parse_solution_to_unitblocks(solution_file)
+    # transformation.parse_txt_to_unitblocks(output_file)
+    transformation.inverse_transformation(nd.n)
 
-###############################################################################################
-##################### Inverse transformation (UCBlock only for now) ###########################
-###############################################################################################
+    differences = compare_networks(network, nd.n)
+    statistics_smspp = nd.n.statistics()
+
+else:
+    ### InvestmentBlock configuration ###
+    configfile = pysmspp.SMSConfig(template="InvestmentBlock/BSPar.txt")
+    temporary_smspp_file = "output/temp_network_investment.nc"
+    output_file = "output/temp_log_file_investment.txt"  # path to the output file (optional)
+    solution_file = "output/temp_solution_file_investment.nc"
+    
+    result = tran.optimize(configfile, temporary_smspp_file, output_file, solution_file, inner_block_name='InvestmentBlock')
+    
+    
+    objective_pypsa = network.objective + network.objective_constant
+    objective_smspp = result.objective_value
+    error = (objective_pypsa - objective_smspp) / objective_pypsa
+    
+    print(f"Error PyPSA-SMS++ of {error}%")
 
 
-# Esegui la funzione sul file di testo
-data_dict = parse_txt_file(output_file)
-
-print(f"Il solver ci ha messo {data_dict['elapsed_time']}s")
-print(f"Il tempo totale (trasformazione+pysmspp+ottimizzazione smspp) è {datetime.now() - then}")
 
 
-solution = transformation.parse_solution_to_unitblocks(solution_file)
-# transformation.parse_txt_to_unitblocks(output_file)
-transformation.inverse_transformation(nd.n)
-
-differences = compare_networks(network, nd.n)
-statistics_smspp = nd.n.statistics()
